requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.7.4
colorama==0.4.6
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
rsa==4.9
Telethon==1.36.0
typing_extensions==4.12.2
yarl==1.9.4
